refactoring #393: moved endpoint to another path
- moved endpoint from DisciplineTeacherController to TeacherController
- moved endpoint service to TeacherService
- changed naming

closes #393
