+++ src/v2/api/services/DisciplineTeacherService.ts
@@ -282,30 +282,4 @@ export class DisciplineTeacherService {
     }
     await this.disciplineTeacherRepository.deleteById(disciplineTeacher.id);
   }
-
-  async getUserDisciplineTeachers (teacherId: string, userId: string, notAnswered: boolean) {
-    return this.disciplineTeacherRepository.findMany({
-      where: {
-        teacherId,
-        discipline: {
-          group: {
-            students: {
-              some: {
-                userId,
-              },
-            },
-          },
-        },
-        ...DatabaseUtils.getOptional(notAnswered, {
-          NOT: {
-            questionAnswers: {
-              some: {
-                userId,
-              },
-            },
-          },
-        }),
-      },
-    });
-  }
 }
